Further componentize DetailsPane.js - Minor refactor/formatting
